requirements.txt
opencv-python
streamlit
mediapipe
numpy
pandas
random
requests
base64
matplotlib
seaborn
pillow
tensorflow
keras
keras-utils
llama-cpp-python
spotipy
os
keras
tensorflow
base64
torch
torchaudio
audiocraft